% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accounts.R
\name{AccountsTree,character-method}
\alias{AccountsTree,character-method}
\title{AccountsTree(yamlstring )}
\usage{
\S4method{AccountsTree}{character}(yamlstring)
}
\arguments{
\item{yamlstring}{character: string describing tree structure see examples}
}
\value{
AccountsTree S4 object: initialized, ready for Financial Model use
}
\description{
This method creates and initializes a new data.tree AccountsTree instance.
Input paramter yamlstring is a character string  in yaml format describing 
the desired AccountsTree structure using indentation for accounts nodes at 
lower levels and listing zero or more ACTUS contractIDs at leaf nodes. The 
tree is created in the $root attribute of the AccountsTree. Its nodes are 
assigned unique nodeIDs. The new AccountsTree objects is returned
}
\examples{
{
yamlstring <- 
"
 name:  modelBank 
 Assets:
   Current:
     actusCIDs:
       - pam001
       - ann002
   ShortTerm:
     actusCIDs:
       - pam003
 Liabilities:
   Debt:
"
accountsTree <- AccountsTree(yamlstring)
}
}
