% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond.R
\name{bond}
\alias{bond}
\title{\code{bond}}
\usage{
bond(start, maturity, nominal, coupon, couponFreq, role)
}
\arguments{
\item{start}{a character reflecting the start date of the bond.}

\item{maturity}{a character reflecting the maturity of the bond,
default is "0 years".}

\item{nominal}{a numeric to set the notional principal of the bond,
default is 0.0.}

\item{coupon}{a numeric to set the coupon payment, default is 0.0.}

\item{couponFreq}{a character reflecting the frequency of coupon payments,
default is "1 year".}

\item{role}{a character reflecting the contract role. default is "long".}
}
\value{
a PrincipalAtMaturity contrat with specified attributes.
}
\description{
Constructor method for simple (single) PrincipalAtMaturity contract.
}
\examples{
b <- bond("2013-12-31", maturity = "5 years", nominal = 50000,
           coupon = 0.02, couponFreq = "1 years")

}
