% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond.R
\name{bondvr}
\alias{bondvr}
\title{\code{bondvr}}
\usage{
bondvr(start, maturity, nominal, coupon, couponFreq, role)
}
\arguments{
\item{start}{a character string yyyy-mm-dd the start date of the bond.}

\item{maturity}{a character string setting the term of the bond.}

\item{nominal}{a numeric to set the notional principal of the bond,}

\item{coupon}{a numeric to set the coupon payment, default is 0.0.}

\item{couponFreq}{a character string with period of coupon payments,}

\item{role}{a character string setting the contract role.}

\item{rateResetFreq}{optional character string setting a period of RateReset}

\item{rateResetSpread}{optional numeric with rate spread for variable rate}
}
\value{
a PrincipalAtMaturity contract with specified attributes.
}
\description{
Constructor for a simple PrincipalAtMaturity fixed or variable rate contract
.
Variable rate bonds  are identified by a nonnull rateResetFrequency. For 
these variable rate bonds the default marketObjectCodeOfRateReset is 
"YC_EA_AAA"
}
\examples{
{
    b <- bondvr("2013-12-31", maturity = "5 years", nominal = 50000,
              coupon = 0.02, couponFreq = "3 months", role = "long",
              rateResetFreq = "1 years", rateResetSpread = 0.01 )
    }
}
