% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContractPlot.R
\name{cashflowPlot,EventSeries-method}
\alias{cashflowPlot,EventSeries-method}
\title{cashflowPlot(<eventSeries> )  - plots a graphical representation od cashflow events}
\usage{
\S4method{cashflowPlot}{EventSeries}(x)
}
\arguments{
\item{x}{The object carrying the contract events to be potted}
}
\value{
creates  returns a graphical canvas (plot) displayed as plot
}
\description{
Creates and displays a graphical representation of the cashflow events of 
  an  \code{\link{EventSeries}} object. An EventSeries captures the cashflows
  of a single ACTUS contract. plot(<eventseries> ) displays the cashflow  
  events of the contract graphically.
}
\details{
The graphical representation shows the different types of cashflow event
  such as Principal Payment, Interest Payment, etc. Cashflows are shown 
  as directional arrows. Incoming cashflows appear as arrows towards the 
  x-axis; outgoing cashflows appear as arrows away from the x-axis. 
  A color code indicates the type of the cashflow as described in detail in 
  the displayed legend of the plot.
}
\examples{
{
  pam1 <- bond("2013-12-31", maturity = "5 years", nominal = 50000,
               coupon = 0.02, couponFreq = "1 years")
  serverURL <- "https://demo.actusfrf.org:8080/" 
  evs1 <- EventSeries(pam1, list(), serverURL)
  cashflowPlot(evs1)     
}
}
