% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContractPlot.R
\name{cashflowPlot}
\alias{cashflowPlot}
\title{cashflowPlot < >     generic function to plot cashflow events of a contract}
\usage{
cashflowPlot(evs)
}
\arguments{
\item{evs}{S4 reference to \code{\link{EventSeries}} object}
}
\description{
The instance of this function is cashflowPlot(\code{\link{EventSeries}}) 
 where the input \code{\link{EventSeries}} carries cashflow event data for a 
 contract to be plotted
}
