% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContractABC.R
\docType{class}
\name{ContractABC-class}
\alias{ContractABC-class}
\title{A Reference Class parent to all ACTUS & FEMS Contract Types}
\description{
This class is only used in an abstract sense in that it is
not intended to create instances of it. It's use is to
serve as parent to various implementations of ACTUS and FEMS CTs
as means for designing method-inheritage.
}

