% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContractModel.R
\docType{class}
\name{ContractModel-class}
\alias{ContractModel-class}
\title{A Reference Class that represents all Contract Models objects}
\description{
This class is only used in a virtual sense in that it is
not intended to create instances of it. It's use is to
provide the class structure, i.e. field definitions, to
various implementations of Contract Models for specific
Contract Types and as a means for designing method-inheritage.
}

