% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContractType.R
\docType{class}
\name{ContractType-class}
\alias{ContractType-class}
\title{ContractType is  Reference Class parent to all ACTUS Contract Types}
\description{
This class is only used in an abstract sense in that it is
not intended to create instances of it. It's use is to
serve as parent to various implementations of ACTUS CTs as
a means for designing method-inheritance.'
}

