% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoSteps.R
\name{createPortfolioFromExcelData}
\alias{createPortfolioFromExcelData}
\title{createPortfolioFromExcelData()}
\usage{
createPortfolioFromExcelData()
}
\description{
reads contract data and risk factor data from CSV files to create
and return a portfolio of ACTUS contract
}
