% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinancialModel.R
\name{events2dfByPeriod,FinancialModel,ANY-method}
\alias{events2dfByPeriod,FinancialModel,ANY-method}
\title{events2dfByPeriod(host = <FinancialModel>)}
\usage{
\S4method{events2dfByPeriod}{FinancialModel,ANY}(host)
}
\arguments{
\item{host}{FinancialModel S4 obj with currentScenarioAnalysis and Timeline}
}
\value{
log msg reporting success of cashflow event bucketing
}
\description{
This method reorganizes the cashflow events in the currentSecnarioAnalysis 
  of the FinancialModel by period using the Timeline of the FInancialModel 
  into a data frame with columns for: contractID, period, and for each 
  ACTUS cashflow event field. The input Financial Model must have: (1)
  a defined portfolio and timeline with status date of all comntracts in the 
  portfolio matching the statusDate in the timeline (2) a defined 
  currentScenarioAnlysis (3) generateEvents( ) must have run successfully on 
  the FinancialModel ( using that current Scenario Analysis ). You can check 
  this using: 
 > unlist(lapply(fm$currentScenarioAnlysis$cashflowEventsLoL,
                 function(x){return(x$status)})) 
  If these conditions are met, events2dfByPeriod() will reorganize the data
  in cthe scna$cashflowEventsLoL as a dataframe with columns: 
  and save that as scna$cashflowEventsByPeriod for use in subsequent analysis
  steps in the currentScenarioAnalysis attribute  
  
  A text message is returned reporting on any issues in this processing step.
  
  Processing steps: (0) check valid host$cashflowEventsLoL, (1) merge 
  eventsLOL into eventsDF, (2) add periodIndex column,  (3) sort by 
  (contractID, periodIndex), (4) save as host$cashFlowEventsByPeriod.
}
\examples{
{

}
           
}
