% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScenarioAnalysis.R
\name{events2dfByPeriod,ScenarioAnalysis,Timeline-method}
\alias{events2dfByPeriod,ScenarioAnalysis,Timeline-method}
\title{events2dfByPeriod(host = <ScenarioAnalysis>, tl = <Timeline>)}
\usage{
\S4method{events2dfByPeriod}{ScenarioAnalysis,Timeline}(host, tl)
}
\description{
This method reorganizes a list(by contract) of lists of cashflow events
  into a data frame with columns for: contractID, period, and for each 
  ACTUS cashflow event field. The input ScenarioAnalysis object must have 
  run generateEvents(host = scna) to populate scna$cashflowEventsLoL, and
  the status of each contract simulation must be "Success" . You can check 
  this using: 
 > unlist(lapply(cfla$cashflowEventsLoL,function(x){return(x$status)})) 
 
  If these conditions are met, events2dfByPeriod() will reorganize the data
  in cfla$cashflowEventsLoL as a dataframe with columns: 
  and save that as scna$cashflowEventsByPeriod for use in subsequent analysis
  steps on the ScenarioAnalysis object. 
  
  A text message is returned reporting on any issues in this processing step.
  
  Processing steps: (0) check valid host$cashflowEventsLoL, (1) merge 
  eventsLOL into eventsDF, (2) add periodIndex column,  (3) sort by 
  (contractID, periodIndex), (4) save as host$cashFlowEventsByPeriod.
}
