% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventSeries.R
\name{EventSeries,ANY,ANY,ANY-method}
\alias{EventSeries,ANY,ANY,ANY-method}
\title{EventSeries(...)  is the generic contstructor for EventSeries objectd}
\usage{
\S4method{EventSeries}{ANY,ANY,ANY}()
}
\description{
EventSeries(...)  is the generic contstructor for EventSeries objectd
}
