% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{eventsLoL2DF}
\alias{eventsLoL2DF}
\title{eventsLoL2DF(cfls)}
\usage{
eventsLoL2DF(cfls)
}
\arguments{
\item{cfls}{cashflow events list of lists from generateEvents}
}
\value{
events dataframe
}
\description{
This functions takes as input (1) a List of Lists generated by the generateEvents 
 function. It converts the list of lists into a merged dataframe and appends the 
 contractId to each event row.
}
