% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinancialModel.R
\name{extendedEventsDF}
\alias{extendedEventsDF}
\title{extendedEventsDF("FinancialModel")}
\usage{
extendedEventsDF(fm)
}
\arguments{
\item{Financial}{Model with cashflowEventsByPeriod available}
}
\value{
events data frame with added fields nodeId, nodeName, contractType, periodStart, periodEnd
}
\description{
This function returns a data frame with the cashflow events in the
currentScenarioAnalysis extended with additional fields for nodeId, nodeName,
contractType, periodStart, and periodEnd. The input Financial Model must have
an accountsTree for the mapping of contracts to nodes and a timeline for the
period dates.
}
