% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accounts.R
\name{extractCIDs}
\alias{extractCIDs}
\title{extractCIDs("accounsTree$root")}
\usage{
extractCIDs(node)
}
\arguments{
\item{node}{AccountsTree$root: the root node of the accounts tree}
}
\value{
vector with all contract IDs of actus contracts and functional contracts
}
\description{
extractCIDs("accounsTree$root")
}
