% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinancialModel.R
\name{FinancialModel,character,character,character,Node,Portfolio,character,Timeline,character-method}
\alias{FinancialModel,character,character,character,Node,Portfolio,character,Timeline,character-method}
\title{FinancialModel( < > ) constructor to create a financial model and set 
its pre-analysis attributes
FinancialModel(fmID, fmDescr, entprID, accounts, ptf, curr, timeline, 
               serverURL)}
\usage{
\S4method{FinancialModel}{character,character,character,Node,Portfolio,character,Timeline,character}(
  fmID,
  fmDescr,
  entprID,
  accounts,
  ptf,
  curr,
  timeline,
  serverURL
)
}
\arguments{
\item{fmID}{character: a unique ID for this financial model}

\item{fmDescr}{character: a short text describing the financial model}

\item{entprID}{character: a unique ID for the enterprise being modelled}

\item{accounts}{Node: tree structure, names- account lines in balance sheets}

\item{ptf}{Portfolio: list of enterprise holdings - ACTUS contracts}

\item{curr}{character: currency for all analysis amounts e.g. CHF, EUR, USD}

\item{timeline}{Timeline - sets timing of projected balance sheet reports}

\item{serverURL}{character URL of ACTUS contract simulation server}
}
\value{
FinancialModel S4 object: ready for analyses to be added
}
\description{
FinancialModel( < > ) constructor to create a financial model and set 
its pre-analysis attributes
FinancialModel(fmID, fmDescr, entprID, accounts, ptf, curr, timeline, 
               serverURL)
}
\examples{
{Hold off on examples until tested and sample Enterprise portfolio added}

}
