% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContractAnalysis.R
\name{generateEvents,ContractAnalysis,missing,missing,missing-method}
\alias{generateEvents,ContractAnalysis,missing,missing,missing-method}
\title{generateEvents(<ContractAnalysis>)}
\usage{
\S4method{generateEvents}{ContractAnalysis,missing,missing,missing}(host)
}
\arguments{
\item{cfla}{CashAnalysis S4 object with portfolio, actusServer and risk data}
}
\value{
Log summarizing which contracts were successfully simulated
}
\description{
The generateEvents(ContractAnalysis) function takes as input a S4
  ContractAnalysis object with the following attributes set: (1) the 
  portfolio of contracts held by the enterprise (2) the actusServerURL
  and (3) the scenario / riskFactor list to be used. The function sends a 
  simulation request to the designated ACTUS server with the contract and 
  risk data and saves the results of the simulation in the ContractAnalysis
  The method return a log message with a report on which contracts were 
  successfully simulated
}
\examples{
{
   mydatadir <- "~/mydata"
   installSampleData(mydatadir)
   cdfn  <- "~/mydata/BondPortfolio.csv"
   ptf   <-  samplePortfolio(cdfn)
   serverURL <- "https://dadfir3-app.zhaw.ch/"
   rxdfp <- paste0(mydatadir,"/UST5Y_fallingRates.csv")
   rfx <- sampleReferenceIndex(rxdfp,"UST5Y_fallingRates", "YC_EA_AAA",100)
   cfla <- initContractAnalysis( analysisID = "cfla001", 
                             analysisDescription = "this_analysis_descr",
                             enterpriseID = "entp001", yieldCurve = YieldCurve(),
                             portfolio =  ptf, currency = "USD", 
                             scenario = list(rfx), 
                             actusServerURL = serverURL, 
                             timeline = Timeline())
   logMsgs  <- generateEvents(host = cfla)
}

}
