% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{generateEvents,Portfolio,character,ANY-method}
\alias{generateEvents,Portfolio,character,ANY-method}
\title{generateEvents(<portfolio>, <ACTUS-server-URL>)}
\usage{
\S4method{generateEvents}{Portfolio,character,ANY}(ptf, serverURL)
}
\arguments{
\item{ptf}{Portfolio S4 object initialized with contract and risk factors}

\item{serverURL}{character - identifies the ACTUS server to be called}
}
\value{
List of generated cashflow results - one entry per contract
}
\description{
The generateEvents(Portfolio, character) function takes an initialized
  S4 Portfolio object and a serverURL as inputs, constructs a JSON
  representation of the Portfolio contents and calls out using https POST to
  the Actus server at URL serverURL to generate a list of cashflow event
  lists for each contract in the Portfolio using the portfolio's RiskFactor
  scenario data.
}
\examples{
{
   mydatadir <- "~/mydata"
   installSampleData(mydatadir)
   cdfn  <- "~/mydata/BondPortfolio.csv"
   rfdfn <- "~/mydata/RiskFactors.csv"
   ptf   <-  samplePortfolio(cdfn,rfdfn)
   serverURL <- "https://demo.actusfrf.org:8080/"
   cfls  <- generateEvents(ptf,serverURL)
}

}
