% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{generateEvents,Portfolio,character,list-method}
\alias{generateEvents,Portfolio,character,list-method}
\title{generateEvents(<portfolio>, <ACTUS-server-URL>, list<riskFactors>)}
\usage{
\S4method{generateEvents}{Portfolio,character,list}(ptf, serverURL, riskFactors)
}
\arguments{
\item{ptf}{Portfolio S4 object initialized with contract and risk factors}

\item{serverURL}{character - identifies the ACTUS server to be called}

\item{riskFactors}{list - list of S4 class riskFactor objects}
}
\value{
List of generated cashflow results - one entry per contract
}
\description{
The generateEvents(Portfolio, character, list) function takes as
  input: (1) an initialized S4 Portfolio of ACTUS contracts, (2) an ACTUS 
  serverURL,  and (3)  a list of riskFactors defining the risk scenario; 
  the function constructs a JSON representation of the Portfolio contents and
  risk factors then calls out using https POST to the Actus server at URL 
  serverURL to generate a list of cashflow event lists for each contract in 
  the Portfolio using the provided list of riskFctors as scenario data.
}
\examples{
{
   mydatadir <- "~/mydata"
   installSampleData(mydatadir)
   cdfn  <- "~/mydata/BondPortfolio.csv"
   ptf   <-  samplePortfolio(cdfn)
   serverURL <- "https://demo.actusfrf.org:8080/"
   rxdfp <- paste0(mydatadir,"/UST5Y_fallingRates.csv")
   rfx <- sampleReferenceIndex(rxdfp,"UST5Y_fallingRates", "YC_EA_AAA",100)
   cfls  <- generateEvents(ptf,serverURL,list(rfx))
}

}
