% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScenarioAnalysis.R
\name{generateEvents,ScenarioAnalysis,Portfolio,character,missing-method}
\alias{generateEvents,ScenarioAnalysis,Portfolio,character,missing-method}
\title{generateEvents(ScenarioAnalysis, Portfolio, ServerURL )}
\usage{
\S4method{generateEvents}{ScenarioAnalysis,Portfolio,character,missing}(host, ptf, serverURL)
}
\arguments{
\item{host}{ScenarioAnalysis S4 object with risk scenario data}

\item{ptf}{Portfolio of ACTUS contracts to be simulated}

\item{serverURL}{character string locating ACTUS server to be used}
}
\value{
Log message listing which contracts were successfully simulated
}
\description{
The generateEvents(ScenarioAnalysis) function takes as input: an 
  initialized S4 ScenarioAnalysis object with RiskFactor information, (2) a 
  Portfolio of ACTUS contracts to be simulated and (3) the URL of an ACTUS
  server to compute the cashflow events. The method sends a 
  JSON simulation request to the designated ACTUS server with portfolio and 
  risk data then saves the results of the simulation in the cashflowEventsLoL
  attribute of the ScenarioAnalysis. A log message is saved in the logMsgs
  list attribute with key "generateEvents" and returned as ouput of the 
  method
}
\examples{
{
   mydatadir <- "~/mydata"
   installSampleData(mydatadir)
   cdfn  <- "~/mydata/TestPortfolio.csv"
   ptf   <-  samplePortfolio(cdfn)
   serverURL <- "https://demo.actusfrf.org:8080/"
   rxdfp <- paste0(mydatadir,"/UST5Y_fallingRates.csv")
   rfx <- sampleReferenceIndex(rxdfp,"UST5Y_fallingRates", "YC_EA_AAA",100)
   rfxs <-list(rfx)
   scnID <- "UST5Y_fallingRates"
   yc <- YieldCurve()
   scna <- ScenarioAnalysis(scenarioID= scnID, marketData= rfxs, 
                            yieldCurve = yc)
  logMsgs  <- generateEvents(host= scna, ptf=ptf, serverURL = serverURL)
}

}
