% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventSeries.R
\name{generateEventSeries,ContractType,list,character-method}
\alias{generateEventSeries,ContractType,list,character-method}
\title{generateEventSeries        <contract>, <risk-factor-list>, <ACTUS-server-URL>}
\usage{
\S4method{generateEventSeries}{ContractType,list,character}(contract, riskFactors, serverURL)
}
\arguments{
\item{contract}{S4 ref      class= ContractType}

\item{riskFactors}{list        list of S4 ref Class=RiskFactor}

\item{serverURL}{character   URL of ACTUS server to simulate the contract}
}
\value{
S4 ref     class=EventSeries
}
\description{
exported function to simulate a contract cashflows and create an
  EventSeries using  "ContractType", "list", "character" method instance.
  constructs an EventSeries instance including as its events_df attribute a
  dataframe of cashflow events for the input ACTUS contract. This cashflow
  is generated with a callout to the ACTUS server located at ACTUS-server-URL
  using a risk scenario specified as the list of risk factors. THe method
  works by first creating a Portfolio with this ine contract and the supplied
  risk factor list, then calling generateEvents on this portfolio
}
\details{
This is an internal method on EventSeries, and not exported.
  generateEventSeries <contract>, risk-factors>, <serverURL> is the
  exported wrapper function which is exported from FEMSdevPkg and calls this.
}
