% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{getContract,Portfolio,character-method}
\alias{getContract,Portfolio,character-method}
\title{getContract(ptf, cid)}
\usage{
\S4method{getContract}{Portfolio,character}(ptf, cid)
}
\arguments{
\item{ptf}{S4 ref to class=Portfolio object with list of contracts}

\item{cid}{character - a contractID string to be matched}
}
\value{
EITHER an S4 Ref to a class=ContractType object with this input
                     cid as its its contractID 
              OR NULL if no such match exists
}
\description{
getContract(ptf, cid) takes as input an S4 ref to a Class=Portfolio object
    containing a list of contracts and a character string contractID. The 
    returns either an S4 ref to a Class=ContractType contract object whose 
    contractID the input cid string OR NULL if there is no matching contract
    in the portfolio.
}
\examples{
{
   mydatadir <- "~/mydata"
   installSampleData(mydatadir)
   cdfn  <- "~/mydata/BondPortfolio.csv"
   ptf <- samplePortfolio(cdfn)
   cids <- getContractIDs(ptf)
   cntr <- getContract(ptf, cids[1])
   }

}
