% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{getContract}
\alias{getContract}
\title{getContract <ptf, cid >     Generic method definition}
\usage{
getContract(ptf, cid)
}
\arguments{
\item{ptf}{S4 reference Class=Portfolio Portfolio with a list of contracts}

\item{cid}{A character string contractID.}
}
\value{
An S4 Reference to a portfolio Object if cid is matched or NULL
}
\description{
Defines a generic method on S4 Class Portfolio also taking character string
  contractID as its second input. Returns an S4 reference to an object of 
  Class=ContractType if the input contractID matches a contract in the 
  Portfolio and NULL if it does not.
}
