% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{getContractIDs,Portfolio-method}
\alias{getContractIDs,Portfolio-method}
\title{getContractIDs}
\usage{
\S4method{getContractIDs}{Portfolio}(ptf)
}
\arguments{
\item{ptf}{S4 ref to class=Portfolio object with list of contracts}
}
\value{
Vector of character string contractIDs -  one for each contract
              in the portfolio
}
\description{
getContractIDs(ptf) takes as input an S4 ref to a Class=Portfolio object
    containing a list of contracts. It returns a vector of character string 
    contractID of the contracts in the portfolio.
}
\examples{
{
   mydatadir <- "~/mydata"
   installSampleData(mydatadir)
   cdfn  <- "~/mydata/BondPortfolio.csv"
   rfdfn <- "~/mydata/RiskFactors.csv"
   ptf <- samplePortfolio(cdfn,rfdfn)
   cids <- getContractIDs(ptf)
   }

}
