% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{getContractIDs}
\alias{getContractIDs}
\title{getContractIDs  <ptf>     Generic method definition}
\usage{
getContractIDs(ptf)
}
\arguments{
\item{ptf}{S4 reference Class=Portfolio Portfolio with a list of contracts.}
}
\value{
A vector of character string contractIDs
}
\description{
Defines a generic method on S4 Class Portfolio. Returns a vector with the 
contractIDs of all the contracts in the portfolio.
}
