% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinancialModel.R
\name{getLQreports}
\alias{getLQreports}
\title{getLQreports("FinancialModel")}
\usage{
getLQreports(fm, scale = 1, rounding = 0)
}
\arguments{
\item{Financial}{Model with accountLQreports() available}
}
\value{
matrix with account Liquidity reports
}
\description{
This function returns a matrix of doubles showing expected nominal value 
reports on liquidity at different dates for the accounts in the financial 
model with #' cashflows generated using the risk environment of the 
currentScenarioAnalysis #' in the financial model. There is a row in the data
frame for each account in the financial model Accounts tree.
}
