% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinancialModel.R
\name{getNMVreports}
\alias{getNMVreports}
\title{getNMVreports("FinancialModel")}
\usage{
getNMVreports(fm, scale = 1, rounding = 0)
}
\arguments{
\item{scale}{numeric: factor to scale nominal values by}

\item{rounding}{numeric: number of decimal places to round to}

\item{Financial}{Model with accountNominalValues() available}
}
\value{
matrix with account Nominal Value reports
}
\description{
This function returns a matrix of doubles showing expected nominal value 
reports at different dates for the accounts in the financial model with 
cashflows generated using the risk environment of the currentScenarioAnalysis
in the financial model. There is a row in the data frame for each account in
the financial model Accounts tree.
}
