% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{installSampleCode}
\alias{installSampleCode}
\title{installSampleCode}
\usage{
installSampleCode(demodir)
}
\arguments{
\item{demodir}{character - full path name of folder to write demo R code}
}
\description{
This function copies the code of a sample R application importing and using
 the FEMSdevPkg package in a user specified directory. The copied code
 introductoryDemo.R provides a simple guide and introduction to using the
 package. This code is documented with comments  introductoryDemo.R is copied.
}
\details{
The input parameter <directory-file-path> of installSampleCode locates a
 directory where the sample data files should be copied to.
}
\examples{
{
  demodir <- "~/mycode"
  installSampleCode(demodir)
  }

}
