% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleContract.R
\name{loan}
\alias{loan}
\title{\code{loan}}
\usage{
{
        ctr1<-  loan(ctype, start, maturity, nominal, coupon, paymentFreq,
                     role, rateResetFreq, rateResetSpread )
       }
}
\arguments{
\item{ctype}{character eith ANN or PAM}

\item{start}{character string yyyy-mm-dd the start date of the mortgage.}

\item{maturity}{character string setting the lifetime of the mortgage.}

\item{nominal}{numeric to set the notional principal of the bond,}

\item{coupon}{numeric initial interest rate 0.02 = 2 pa default is 0.0.}

\item{paymentFreq}{character string period of payments (interest + principal)}

\item{role}{character string setting whether lender or borrower role.}

\item{rateResetFreq}{optional character string setting a period of RateReset}

\item{rateResetSpread}{optional numeric with rate spread for variable rate}
}
\value{
initialized Annuity contract contract with specified attributes.
}
\description{
Function loan() creates and returns a PAM or an ANN
}
\examples{
{
    l <- loan("ANN", "2020-12-31", maturity = "10 years", nominal = 10000,
              coupon = 0.07, paymentFreq = "3 months", role = "long",
              rateResetFreq = "1 year", rateResetSpread = 0.01 )
    }
}
