% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContractAnalysis.R
\name{lv2LiquidityReports,ContractAnalysis-method}
\alias{lv2LiquidityReports,ContractAnalysis-method}
\title{*******************************
  lv2LiquidityReports(cfla= ContractAnalysis) -method instance
*******************************}
\usage{
\S4method{lv2LiquidityReports}{ContractAnalysis}(host)
}
\arguments{
\item{host}{CashAnalysis S4 object with portfolio, actusServer and risk data}
}
\value{
Log summarizing whether processing was successful
}
\description{
This method generates a liquidityReport (vector) for each contract using 
  the data in cfla$contractLiquidityVectors and saves this as a list keyed by 
  contractID in cfla$liquidityReports. The method is exported
  
  Differences between liquidityReports and contractLiquidityVectors are:
  (1) a contractLiquidityVector can have up to cfla$timeline$periodCount 
  elements, (2) but it actually only has entries for periods in which at least
  one liquidity changing event happens ( so it can also have fewer than 
  cfla$timeline$reportCount entries (3) a liquidityReport vector must have 
  exactly reportCount elements (4) contractLiquidityVectors record the 
  change in liquidity for each period in which the related contract sees
  liquidity changing events (5) It is more convenient for liquidity
  reporting to state the cumulative change in enterprise liquidity caused by
  each contract at each report date. That way, by aggregating the 
  liquidIty reports for all contracts, We can immediately see if the 
  enterprise is solvent at each report date. 
  
  The method returns a message indicating whether processing was successful.
}
\examples{
{
   mydatadir <- "~/mydata"
   installSampleData(mydatadir)
   cdfn  <- "~/mydata/BondPortfolio.csv"
   ptf   <-  samplePortfolio(cdfn)
   ptfsd <- unlist(lapply(ptf$contracts,function(x){return(x$contractTerms["statusDate"])}))
   ptf2015 <- Portfolio(contractList = ptf$contracts[which(ptfsd == "2015-01-01")])
   serverURL <- "https://dadfir3-app.zhaw.ch/"
   rxdfp <- paste0(mydatadir,"/UST5Y_fallingRates.csv")
   rfx <- sampleReferenceIndex(rxdfp,"UST5Y_fallingRates", "YC_EA_AAA",100)
   tl1 <- Timeline("2015-01-01",3,4,8)
   cfla2015 <- initContractAnalysis( analysisID = "cfla001", 
                             analysisDescription = "this_analysis_descr",
                             enterpriseID = "entp001", yieldCurve = YieldCurve(),
                             portfolio =  ptf2015, currency = "USD", 
                             scenario = list(rfx), 
                             actusServerURL = serverURL, 
                             timeline = tl1)
   logMsgs1  <- generateEvents(host = cfla2015)
   logMsgs2  <- events2dfByPeriod(host = cfla2015)
   logMsgs3  <- liquidityByPeriod2vec(host= cfla2015)
   lofMsgs4  <- lv2LiquidityReports(host= cfla2015)
}      
}
