% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{monthlyAverageRate}
\alias{monthlyAverageRate}
\title{monthlyAverageRate indf)  function}
\usage{
monthlyAverageRate(indf)
}
\arguments{
\item{indf}{input dataframe with columns named Rate and Date}
}
\value{
outdf  Rate values from indf aggregated by month using mean()
}
\description{
This function takes as input a data frame with a numeric Rate column
   and a charcter string Date columns. It returns as output a dataframe
   with  a numeric column Rate and a timeDateCOlumn Date  - but the Rate
   values now aggregated using
   FUN = mean and each months data associated with the first day of the month.
}
\examples{
{
 indf <- data.frame(Date = as.Date(c('2001-01-01','2001-01-08', '2001-01-15',
                        '2001-02-01')),
                  Rate = c( 3.1, 3.2, 3.3, 4.1)
                  )
 outdf <- monthlyAverageRate(indf)
}

}
