% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleContract.R
\name{mortgage}
\alias{mortgage}
\title{\code{mortgage}}
\usage{
mortgage(start, maturity, nominal, coupon, paymentFreq, role, 
                     rateResetFreq, rateResetSpread )
}
\arguments{
\item{start}{character string yyyy-mm-dd the start date of the mortgage.}

\item{maturity}{character string setting the lifetime of the mortgage.}

\item{nominal}{numeric to set the notional principal of the bond,}

\item{coupon}{numeric to set the initial interest rate; default is 0.0.}

\item{paymentFreq}{character string period of payments (interest + principal)}

\item{role}{character string setting whether lender or borrower role.}

\item{rateResetFreq}{optional character string setting a period of RateReset}

\item{rateResetSpread}{optional numeric with rate spread for variable rate}
}
\value{
initialized Annuity contract contract with specified attributes.
}
\description{
mortgage()   function to construct a single annuity contract using 
             simplified parameters. Handles fixed and variable rates. 
             Variable rate is identified by a nonnull rateResetFreq. The 
             market object code for rate reset is by default the interest 
             market object code YC_EA_AAA
}
\examples{
{
    m <- mortgage("2020-12-31", maturity = "5 years", nominal = 50000,
              coupon = 0.02, paymentFreq = "3 months", role = "long",
              rateResetFreq = "1 year", rateResetSpread = 0.01 )
    }
}
