% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{Portfolio,ANY-method}
\alias{Portfolio,ANY-method}
\title{Portfolio ( )  - no parameters instance of Portfolio< >}
\usage{
\S4method{Portfolio}{ANY}()
}
\value{
S4 reference with class=Portfolio and no attributes initialized.
}
\description{
Creates an empty Portfolio object with no attributes initialized.
}
