% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{Portfolio}
\alias{Portfolio}
\title{Portfolio < >  -  generic function definition}
\usage{
Portfolio(contract, ...)
}
\arguments{
\item{contract}{S4 reference Class=ContractType, a contract to include.}

\item{...}{Not used}
}
\description{
Defines generic S4 constructor method on class Portfolio
}
