% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiskFactor.R
\docType{class}
\name{RiskFactor-class}
\alias{RiskFactor-class}
\title{Class RiskFactor}
\value{
S4 object class RiskFactor
}
\description{
This class is only used in a virtual sense in that it is
 not intended to create instances of it. It's use is to
 serve as parent to various implementations of Risk Factor
 types such as  MarketIndex, YieldCurve .
}
\section{Fields}{

\describe{
\item{\code{label}}{character.}

\item{\code{base}}{numeric.}

\item{\code{data}}{timeSeries.}
}}


