% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyDemo.R
\name{runDaDFiR3demo}
\alias{runDaDFiR3demo}
\title{runDaDFiR3demo()  Launches a shiny demo application}
\usage{
runDaDFiR3demo()
}
\description{
Launches a browser controlled application plotting the provided sample 
reference rate data in a display window. You can chose between plots 
of  5yTreasuryBill_fallingRates, 5yTreasuryBill_recoveringRates, etc
}
