% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{samplePortfolio}
\alias{samplePortfolio}
\title{samplePortFolio}
\usage{
samplePortfolio(cdfn, rfdfn)
}
\arguments{
\item{cdfn}{character string -  a contract-data-filepath}

\item{rfdfn}{character string -  a riskfactor-data-filepath}
}
\value{
Portfolio s4 object initialized with the data from the input files
}
\description{
samplePortfolio ( cdfn, rdfn ) takes as input a contracts-data-filepath and
    riskfactor- data-filepath, reads this data and returns an initialized
    Portfolio object with contracts and risk factors from these csv files.
}
\examples{
{
   mydatadir <- "~/mydata"
   installSampleData(mydatadir)
   cdfn  <- "~/mydata/BondPortfolio.csv"
   rfdfn <- "~/mydata/RiskFactors.csv"
   ptf <- samplePortfolio(cdfn,rfdfn)
   }

}
