% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReferenceIndex.R
\name{sampleReferenceIndex}
\alias{sampleReferenceIndex}
\title{sampleReferenceIndex     (rxdfp,rfID, moc, base)
 
    Function to read a csv file with a column of yyyy-mm-dd dates, and a 
    column of interest rate values. A marketObjectCode, a riskFactorID and 
    a numeric base value. The function returns an S4 ref to an object with
    class=ReferenceIndex. The marketObjectCode is the identifier used in 
    variable rate contracts to specify a dependency on this ReferenceIndex
    for setting a nominal interest rate.  The numeric value base should have 
    value 100 if a 3.1% interest rate appears as 3.1 in the csv file, and 
    value 1.0 if that rate appears as 0.031. Input parameter rfID will be 
    set as the unique riskFactorID of the returned ReferenceIndex object.
    Examples of csv files formatted with Dates and Values for ReferenceIndex
    creation in the sample data are: UST5Y_fallingRates.csv,
    UST5Y_recoveringRates.csv, UST5Y_risingRates.csv, UST5Y_steadyRates.csv .}
\usage{
sampleReferenceIndex(rxdfp, rfID, moc, base)
}
\arguments{
\item{rxdfp}{character  Pathname of csv data file for the referenceIndex}

\item{rfID}{character  the (unique) riskFactorID for new referenceIndex}

\item{moc}{character  marketObjectCode used in varying rate contracts}

\item{base}{numeric    1.0 or 100  if 3 percent  as 0.03 or 3.00}
}
\value{
S4 ref to a new class=ReferenceIndex object
}
\description{
sampleReferenceIndex     (rxdfp,rfID, moc, base)
 
    Function to read a csv file with a column of yyyy-mm-dd dates, and a 
    column of interest rate values. A marketObjectCode, a riskFactorID and 
    a numeric base value. The function returns an S4 ref to an object with
    class=ReferenceIndex. The marketObjectCode is the identifier used in 
    variable rate contracts to specify a dependency on this ReferenceIndex
    for setting a nominal interest rate.  The numeric value base should have 
    value 100 if a 3.1% interest rate appears as 3.1 in the csv file, and 
    value 1.0 if that rate appears as 0.031. Input parameter rfID will be 
    set as the unique riskFactorID of the returned ReferenceIndex object.
    Examples of csv files formatted with Dates and Values for ReferenceIndex
    creation in the sample data are: UST5Y_fallingRates.csv,
    UST5Y_recoveringRates.csv, UST5Y_risingRates.csv, UST5Y_steadyRates.csv .
}
\examples{
{
   mydatadir <- "~/mydata"
   installSampleData(mydatadir)
   rxdfp <- paste0(mydatadir,"/UST5Y_fallingRates.csv")
   rfx <- sampleReferenceIndex(rxdfp,"UST5Y_fallingRates", "YC_EA_AAA",100)
}
}
