% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScenarioAnalysis.R
\name{ScenarioAnalysis,character,list,YieldCurve-method}
\alias{ScenarioAnalysis,character,list,YieldCurve-method}
\title{ScenarioAnalysis (required initialization parameters)   
internal use only 
Creates and returns S4 ref ScenarioAnalysis instance ready for contract 
simulationized.}
\usage{
\S4method{ScenarioAnalysis}{character,list,YieldCurve}(scenarioID, marketData, yieldCurve)
}
\arguments{
\item{scenarioID}{character ID for this risk Scenario}

\item{marketData}{list of risk factor reference indexes}

\item{yieldCurve}{a YieldCurve - related to factors for discounting}
}
\value{
initialized ScenarioAnalysis object
}
\description{
ScenarioAnalysis (required initialization parameters)   
internal use only 
Creates and returns S4 ref ScenarioAnalysis instance ready for contract 
simulationized.
}
