% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialize.R
\name{setup}
\alias{setup}
\title{setup}
\usage{
setup()
}
\value{
A reference to S4 Class Environment ACTUS server url set
}
\description{
initializes the FEMSdevPgk environment sets the ACTUS server url
}
\details{
initialize object env <- setup()
functions in the package will use env
inspect the code in Initialize.r::setup change for different ACTUS url
}
