% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinancialModel.R
\name{showContractNMVs}
\alias{showContractNMVs}
\title{showContractNMVs("FinancialModel")}
\usage{
showContractNMVs(fm, scale = 1, rounding = 0)
}
\arguments{
\item{Financial}{Model with nominalValueReports available}
}
\value{
data frame suitable for displaying results
}
\description{
This function returns a dataframe showing expected nominal value reports at 
different dates for the contracts in the contracts in the financial model
portfolio with cashflows generated using the risk environment of the 
currentScenarioAnalysis. There is a row in the data frame for each contract
and coluns for each report date 
displayed in the first column of the data frame
}
