% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinancialModel.R
\name{showNMVreports}
\alias{showNMVreports}
\title{showNMVreports("FinancialModel")}
\usage{
showNMVreports(fm, scale = 1, rounding = 0)
}
\arguments{
\item{scale}{numeric: factor to scale nominal values by}

\item{rounding}{numeric: number of decimal places to round to}

\item{Financial}{Model with accountNominalValues() available}
}
\value{
data frame suitable for displaying results
}
\description{
This function returns a dataframe showing expected nominal value reports at 
different dates for the accounts in the financial model with cashflows 
generated using the risk environment of the currentScenarioAnalysis in the 
financial model. There is a row in the data frame for #' each account in the 
financial model Accounts tree. The structure of the accounts tree is 
displayed in the first column of the data frame
}
