% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{simulatePortfolio}
\alias{simulatePortfolio}
\title{simulatePortfolio (ptf, serverURL, riskFactors, scenarioName)}
\usage{
simulatePortfolio(ptf, serverURL, riskFactors, scenarioName)
}
\arguments{
\item{ptf}{Portfolio of ACTUS contracts to be simulated class=Portfolio}

\item{serverURL}{locates ACTUS server to generate cashflow events}

\item{riskFactors}{list of riskFactors class RiskFactor - the scenario}

\item{scenarioName}{character   name for the scenario - used in plot title}
}
\value{
vector of plots: income and liquidity change, monthly and cumulative
}
\description{
This functions takes as input (1) a portfolio of ACTUS contracts (2) a list
 of risk factors - the historical and projected future values for interest 
 rates, etc  (3) a URL identifying an ACTUS server to generate cashflow 
 events for the contracts and (4) a scenario name. The ACTUS server is
 invoked via an http POST with all required data for the cashflow simulation
 passes as JSON.
 
 If the cashflow simulation is successful, the returned cashflow events are
 merged into a dataframe sorted by time. Income events and Net capital and 
 interest flows are extracted and  aggregated into monthly time buckets. 
 Function ggplot is used to generate graphics for Interest Income by Month,
 Cumulated Income Month by Month, Liquidity Change By Month, and Cumulative 
 Liquidity Position. A vector with these four plots is returned. 
 
 There may also be riskFactors in the portfolio ptf but they are not used
}
\examples{
{
   mydatadir <- "~/mydata"
   installSampleData(mydatadir)
   cdfn  <- "~/mydata/BondPortfolio.csv"
   ptf <- samplePortfolio(cdfn)
   falling_fp <- paste0(mydatadir,"/UST5Y_fallingRates.csv")
   rfx_falling <- sampleReferenceIndex(falling_fp,"UST5Y_fallingRates", 
                                   "YC_EA_AAA",100)
   serverURL <- "http://ractus.ch:8080/"
   plotlist <- simulatePortfolio(ptf, serverURL, list(rfx_falling),
                               rfx_falling$riskFactorID ) 
   plotlist[["monthly income"]]                               
}
}
