% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Timeline.R
\name{Timeline,character,numeric,numeric,numeric-method}
\alias{Timeline,character,numeric,numeric,numeric-method}
\title{Timeline(statusDate, monthsPerBucket, reportCount, periodCount)}
\usage{
\S4method{Timeline}{character,numeric,numeric,numeric}(statusDate, monthsPerPeriod, reportCount, periodCount)
}
\arguments{
\item{statusDate}{character yyyy-mm-01 date=statusdate of portfolio contracts}

\item{monthsPerPeriod}{integer length in months of period between reports}

\item{reportCount}{integer number of reports(income, liquidity) to produce}

\item{periodCount}{integer number of periods including valuation periods}
}
\value{
Timeline S4 object - initialized, ready to use inCashflowAnalysis()
}
\description{
This method is used to create and initialize a Timeline object which 
  specifies: (1) the statusDate on which current state of enterprise holdings
  is reported. For consistency this should also be the statusDate value 
  specified for all ACTUS contracts in the CashflowAnalysis portlfolio; 
  statusDate also needs to be a "first of month" date e.g. 2024-02-01, 
  2024-03-01 - to simplify the mapping of cashflow event dates into periods. 
  (2) monthsPerPeriod is an integer specifying the number or months between 
  consecutive analysis reports; usually the monthsPerPeriod will be a divisor
  or a multiple of 12 so that generated analysis reports are for the same 
  months in each year.  (3) reportCount specifies how many reports  on
  liquidity, income and present value should be generated for each contract
  (4) periodCount specifies the total number of periods into which the 
  cashflow events for each contract will be grouped and eventually 
  aggregated. We expect periodCount to be appox 2x reportCount. Valuation of
  contract cashflow events occurring beyond the period "horizon" is handled
  without aggregation - in a single open ended final period with index 999.  
  
  The initialized Timeline object is used as an input to CashflowAnalysis
  where it controls the timing of analysis reports, and valuation of far 
  future cashflows. 
  
  Example Timeline(="2024-01-01",3,4,8) creates a timeline for a cashflow 
  portfolio with contracts having  statusDate="2024-01-01". Reports will be
  quarterly. Four reports on liquidity and income will be generated, dated
  2024-04-01, 2024-07-01, 2024-10-01, 2025-01-01. Cashflow events occuring on
  or after 2024-01-01 but before 2024-04-01 will be considered period 1 
  are included in the 2024-04-01 report. Periods 2,3, and 4 are defined
  similarly. Cashflow event occuring on or after "2026-01-01" are beyond the
  period horizon and will be discounted and valued individually ( rather 
  than aggregated into a period net flow before discounting). These events 
  have periodIndex=999. Valuations will be generated for the 
  timeline statusDate 2024-01-01 and for the four subsequent report times.
}
\examples{
{
tl<-Timeline("2024-01-01", 3, 4, 8)
}
}
