% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accounts.R
\name{treeToYamlString}
\alias{treeToYamlString}
\title{treeToYamlString("accounsTree")}
\usage{
treeToYamlString(accountsTree)
}
\arguments{
\item{accountsTree}{AccountsTree: the accounts tree to be converted to yaml}
}
\value{
yaml string: describing the accounts tree structure
}
\description{
This function returns a yaml string representation of the accounts tree, it can later
be used to recreate the accounts tree using the AccountsTree constructor.
}
