% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YieldCurve.R
\name{YieldCurve,ANY,ANY,ANY,ANY,ANY-method}
\alias{YieldCurve,ANY,ANY,ANY,ANY,ANY-method}
\title{YiedCurve ( )  - no parameters instance of YieldCurve}
\usage{
\S4method{YieldCurve}{ANY,ANY,ANY,ANY,ANY}()
}
\value{
S4 reference with class=YieldCurve and no attributes initialized.
}
\description{
Creates an empty YieldCurve with no attributes initialized.
}
