% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YieldCurve.R
\name{YieldCurve,character,character,numeric,character,character-method}
\alias{YieldCurve,character,character,numeric,character,character-method}
\title{YieldCurve(yieldCurveID, referenceDate, tenorRates, dayCountConvention,
            compoundingFrequency )}
\usage{
\S4method{YieldCurve}{character,character,numeric,character,character}(
  yieldCurveID,
  referenceDate,
  tenorRates,
  dayCountConvention,
  compoundingFrequency
)
}
\arguments{
\item{yieldCurveID}{character  label uniquely identifying this yieldCurve}

\item{referenceDate}{character date yyyy-mm-dd tenorRates observed this day}

\item{tenorRates}{numeric  pa rates (0.02=2pc) vector, names "1M", "2Y" etc}

\item{dayCountConvention}{character: "30E360","30E360ISDA","A360","A365","AA"}

\item{compoundingFrequency}{character: "NONE", "YEARLY", "CONTINUOUS"}
}
\value{
fully initialized S4 class YieldCurve object
}
\description{
YieldCurve(character, character, namedNumericVector, character,character)
  function takes as input: (1) a YieldCurveID label, (2) a referenceDate
  character string, (3) a named vector of numeric tenorRates, (4) a
  character string ACTUS dayCountConvention and (5) a character string
  compoundingFrequency with values in { "NONE", "YEARLY","CONTINUOUS"}.
  The function checks that the input dayCountConvention is a valid ACTUS code
  and mappable to the dayCountConvention values supported in
  RQuantLib::yearFraction - used to compute arbitrage free forward rates.
  An S4 YieldCurve object is created and returned with attributes initialized
  to these values.
}
\examples{
{
   ycID <- "yc001"
   rd <- "2023-10-31"
   tr <-  c(1.1, 2.0, 3.5 )/100
   names(tr) <- c("1M", "1Y", "5Y")
   dcc <- "30E360"
   cf <- "NONE"
   yc <- YieldCurve(ycID,rd,tr,dcc,cf)
}
}
