% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YieldCurve.R
\docType{class}
\name{YieldCurve-class}
\alias{YieldCurve-class}
\title{class YieldCurve}
\description{
A YieldCurve object holds data for an (interest rate) yield curve for 
some risk class (most likely riskfree) at a particular reference date. 
This data is organized as a set of spot rates for user selected tenors 
at the time of the reference date.
}
\section{Fields}{

\describe{
\item{\code{yieldCurveID}}{character label identifying this yieldcurve object}

\item{\code{referenceDate}}{character yyyy-mm-dd date for when YC valid}

\item{\code{tenorRates}}{numeric vector spot rates on refdate; rates are per annum 
i.e. 2% pa = 0.02    Each rated labelled with tenor 
rates$names = "1D" "1W" "1M" "3M" "6M" "1Y" "2Y" "5Y"}

\item{\code{tenorYfs}}{numeric vector of tenor names converted to year fractions}

\item{\code{dayCountConvention}}{character ACTUS string eg "30E360"}

\item{\code{yfdcc}}{character  mapped dayCountConvention used in fmdates:year_frac}

\item{\code{compoundingFrequency}}{character "NONE", "YEARLY", "CONTINUOUS"}
}}


