% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleContract.R
\name{bondvr}
\alias{bondvr}
\title{\code{bondvr}
    Function bondvr() creates and returns a fixed or variable rate PAM}
\usage{
bondvr(start, maturity, nominal, coupon, paymentFreq, role, 
              rateResetFreq, rateResetSpread )
}
\arguments{
\item{start}{character string yyyy-mm-dd the start date of the bond.}

\item{maturity}{character string (lubridate) period for term of the bond.}

\item{nominal}{numeric to set the notional principal of the bond,}

\item{coupon}{numeric the coupon/interest rate 0.02 = 2% pa default is 0.0.}

\item{paymentFreq}{a character string with period of coupon payments,}

\item{role}{a character string setting the contract role.}

\item{rateResetFreq}{optional character string setting a period of RateReset}

\item{rateResetSpread}{optional numeric with rate spread for variable rate}
}
\value{
a PrincipalAtMaturity contract with specified attributes.
}
\description{
Convenience function with a simplified parameter list to create and 
    initialize a single PrincipalAtMaturity fixed or variable rate bond
    contract.
.
    Variable rate bonds  are identified by a nonnull rateResetFrequency. For 
    these variable rate bonds the default marketObjectCodeOfRateReset is 
    "YC_EA_AAA"
    
    Terms statusDate and contractDealDate are set 1 day before start date. 
    Term initialExchangeDate is set to start date. The cycleAnchorDates for
    Interest and principal payments are set at ( IED + a period ). The 
    maturityDate is set ( IED + Maturity period) defaults are set for 
    currency, dayCountConvention, Calendar. Role, notionaPrincipal and 
    nominalInterestRate are set as specified in the parameter list.
}
\examples{
{
    b <- bondvr("2013-12-31", maturity = "5 years", nominal = 50000,
              coupon = 0.02, couponFreq = "3 months", role = "long",
              rateResetFreq = "1 years", rateResetSpread = 0.01 )
    }
}
