% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContractPlot.R
\name{cashflowPlot,EventSeries-method}
\alias{cashflowPlot,EventSeries-method}
\title{cashflowPlot  (evs) - create graphical plot of a contract's cashflow events}
\usage{
\S4method{cashflowPlot}{EventSeries}(evs)
}
\arguments{
\item{evs}{a \code{\link{EventSeries}} object with contract events to be 
plotted}
}
\value{
creates  returns a graphical canvas (plot) displayed as plot
}
\description{
Creates and displays a graphical representation of the cashflow events in 
  an  \code{\link{EventSeries}} object. An \code{\link{EventSeries}} captures
  the cashflows of a single ACTUS contract. The exported function
  cashflowPlot(\code{\link{EventSeries}}) displays the cashflow events of the
  contract graphically.
}
\details{
The graphical representation shows the different types of cashflow event
  such as Principal Payment, Interest Payment, etc. Cashflows are shown 
  as directional arrows. Incoming cashflows appear as arrows towards the 
  x-axis; outgoing cashflows appear as arrows away from the x-axis. 
  A color code indicates the type of the cashflow as described in detail in 
  the displayed legend of the plot.
}
\examples{
{
  pam1      <- bond("2013-12-31", maturity = "5 years", nominal = 50000,
                     coupon = 0.02, couponFreq = "1 years", role = "long")
  serverURL <- "https://demo.actusfrf.org:8080/" 
  evs1      <- generateEventSeries(pam1, list(), serverURL)
  cashflowPlot(evs1)     
}
}
