% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleContract.R
\name{mortgage}
\alias{mortgage}
\title{\code{mortgage}}
\usage{
mortgage(start, maturity, nominal, coupon, paymentFreq, role, 
                 rateResetFreq, rateResetSpread )
}
\arguments{
\item{start}{character string yyyy-mm-dd the start date of the mortgage.}

\item{maturity}{character string setting the lifetime of the mortgage.}

\item{nominal}{numeric to set the notional principal of the bond,}

\item{coupon}{numeric initial interest rate 0.02 = 2pcpa default is 0.0.}

\item{paymentFreq}{character string period of payments (interest + principal)}

\item{role}{character string setting whether lender or borrower role.}

\item{rateResetFreq}{optional character string setting a period of RateReset}

\item{rateResetSpread}{optional numeric with rate spread for variable rate}
}
\value{
initialized Annuity contract contract with specified attributes.
}
\description{
Function mortgage() creates and returns a fixed or variable rate ANN
}
\details{
Convenience function with a simplified parameter list to create and 
    initialize a single Annuity fixed or variable rate mortgage contract.
.
    Variable rate bonds  are identified by a nonnull rateResetFrequency. For 
    these variable rate bonds the default marketObjectCodeOfRateReset is 
    "YC_EA_AAA"
    
    Terms statusDate and contractDealDate are set 1 day before start date. 
    Term initialExchangeDate is set to start date. The cycleAnchorDates for
    Interest and principal payments are set at ( IED + a period ). The 
    maturityDate is set ( IED + Maturity period) defaults are set for 
    currency, dayCountConvention, Calendar. Role, notionaPrincipal and 
    nominalInterestRate are set as specified in the parameter list.
}
\examples{
{
    m <- mortgage("2020-12-31", maturity = "10 years", nominal = 10000,
              coupon = 0.07, paymentFreq = "3 months", role = "long",
              rateResetFreq = "1 year", rateResetSpread = 0.01 )
    }
}
