% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\docType{class}
\name{Portfolio-class}
\alias{Portfolio-class}
\title{class Portfolio}
\description{
A Portfolio consists of a list of contracts. This is one of the required
input parameter for a generateEvents(portfolio, scenario) request
}
\section{Fields}{

\describe{
\item{\code{contracts}}{List of contracts, class=ContractType, in the portfolio.}
}}


